# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Pivot API",
        "version": "0.1.0"
    },
    "basePath": "/",
    "paths": {
        "/open/calculateResult": {
            "post": {
                "tags": [
                    "BFF Election Result"
                ],
                "summary": "Run an election using BFF ballots",
                "operationId": "calculateResult",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "text/plain"
                ],
                "parameters": [
                    {
                        "name": "ballots",
                        "in": "formData",
                        "description": "One or more BFF ballots, one on each line",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "tieBreaker",
                        "in": "formData",
                        "description": "One BFF Ballot for resolving intermediate ties",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "description": "JSON with a key `result` whose value is a BFF election result",
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/elections/{electionId}/candidates": {
            "get": {
                "tags": [
                    "Candidates"
                ],
                "summary": "View candidates for an election",
                "operationId": "candidateIndex",
                "parameters": [
                    {
                        "name": "electionId",
                        "in": "path",
                        "description": "Election to get",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Candidate"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Candidates"
                ],
                "summary": "Add a candidate",
                "operationId": "createCandidate",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "electionId",
                        "in": "path",
                        "description": "Election ID",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "payload",
                        "in": "body",
                        "description": "Candidate to add",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CreateCandidate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/Candidate"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/elections/{electionId}/candidates/{candidateId}": {
            "get": {
                "tags": [
                    "Candidates"
                ],
                "summary": "Get information about a candidate",
                "operationId": "getCandidateById",
                "parameters": [
                    {
                        "name": "electionId",
                        "in": "path",
                        "description": "Election to get",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "candidateId",
                        "in": "path",
                        "description": "Candidate to get",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/Candidate"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Candidates"
                ],
                "summary": "Delete a candidate",
                "operationId": "deleteCandidate",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "electionId",
                        "in": "path",
                        "description": "Election ID",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "candidateId",
                        "in": "path",
                        "description": "Candidate ID to delete",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Candidates"
                ],
                "summary": "Update a candidate",
                "operationId": "updateCandidate",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "electionId",
                        "in": "path",
                        "description": "Election ID",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "payload",
                        "in": "body",
                        "description": "Candidate to update",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CreateCandidate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Candidate": {
            "properties": {
                "id": {
                    "description": "",
                    "type": "string",
                    "example": "123"
                },
                "election_id": {
                    "description": "",
                    "type": "string",
                    "example": "123"
                },
                "name": {
                    "description": "",
                    "type": "string",
                    "example": "Funds for road construction"
                },
                "created_at": {
                    "description": "",
                    "type": "string",
                    "example": "2017-01-01 11:00:00"
                },
                "updated_at": {
                    "description": "",
                    "type": "string",
                    "example": "2017-01-01 11:00:00"
                }
            },
            "type": "object",
            "xml": {
                "name": "Candidate"
            }
        },
...
        "CreateCandidate": {
            "properties": {
                "name": {
                    "description": "",
                    "type": "string",
                    "example": "Highland Avenue"
                }
            },
            "type": "object",
            "xml": {
                "name": "CreateCandidate"
            }
        },
...
    }
}